third_party/blink/web_tests/external/wpt/css/css-display/parsing/display-computed-expected.txt

This is a testharness.js-based test.
Found 28 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Property display value 'run-in'
  assert_true: 'run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'ruby-base'
  assert_true: 'ruby-base' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in flow'
  assert_true: 'run-in flow' is a supported value for display. expected true got false
[FAIL] Property display value 'flow run-in'
  assert_true: 'flow run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'flow-root run-in'
  assert_true: 'flow-root run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in flow-root'
  assert_true: 'run-in flow-root' is a supported value for display. expected true got false
[FAIL] Property display value 'flex run-in'
  assert_true: 'flex run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in flex'
  assert_true: 'run-in flex' is a supported value for display. expected true got false
[FAIL] Property display value 'grid run-in'
  assert_true: 'grid run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in grid'
  assert_true: 'run-in grid' is a supported value for display. expected true got false
[FAIL] Property display value 'table run-in'
  assert_true: 'table run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in table'
  assert_true: 'run-in table' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in ruby'
  assert_true: 'run-in ruby' is a supported value for display. expected true got false
[FAIL] Property display value 'ruby run-in'
  assert_true: 'ruby run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in list-item'
  assert_true: 'run-in list-item' is a supported value for display. expected true got false
[FAIL] Property display value 'list-item run-in'
  assert_true: 'list-item run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'flow run-in list-item'
  assert_true: 'flow run-in list-item' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in flow list-item'
  assert_true: 'run-in flow list-item' is a supported value for display. expected true got false
[FAIL] Property display value 'flow list-item run-in'
  assert_true: 'flow list-item run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in list-item flow'
  assert_true: 'run-in list-item flow' is a supported value for display. expected true got false
[FAIL] Property display value 'list-item run-in flow'
  assert_true: 'list-item run-in flow' is a supported value for display. expected true got false
[FAIL] Property display value 'list-item flow run-in'
  assert_true: 'list-item flow run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'flow-root run-in list-item'
  assert_true: 'flow-root run-in list-item' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in flow-root list-item'
  assert_true: 'run-in flow-root list-item' is a supported value for display. expected true got false
[FAIL] Property display value 'flow-root list-item run-in'
  assert_true: 'flow-root list-item run-in' is a supported value for display. expected true got false
[FAIL] Property display value 'run-in list-item flow-root'
  assert_true: 'run-in list-item flow-root' is a supported value for display. expected true got false
[FAIL] Property display value 'list-item run-in flow-root'
  assert_true: 'list-item run-in flow-root' is a supported value for display. expected true got false
[FAIL] Property display value 'list-item flow-root run-in'
  assert_true: 'list-item flow-root run-in' is a supported value for display. expected true got false
Harness: the test ran to completion.
